Se crea ContactFormComponent para el formulario de contacto y se agrega UseQuery en ContactComponent
