+++ src/App.css
@@ -25,7 +25,6 @@ ul, li {
 
 .container {
   width: 95%;
-  height:100vh;
   margin: auto;
   padding: 20px 0;
   text-align: justify;
