+++ src/App.js
@@ -2,7 +2,7 @@ import './App.css';
 import { BrowserRouter as Router, Route } from 'react-router-dom';
 import { NavbarComponent } from './navbar/navbar-component';
 import { HomeComponent } from './home/home-component';
-import { ContactComponent } from './contact/contact-componet';
+import { ContactComponent } from './contact/contact-component';
 import { ExampleGQLComponent } from './example-gql/example-gql-component';
 import { ScreenErrorComponent } from './screen-error/screen-error-component';
 import { FooterComponent } from './footer/footer-component';
@@ -15,7 +15,6 @@ function App() {
         <Route exact path="/" component={HomeComponent} />
         <Route path="/contact" component={ContactComponent} />
         <Route path="/example-gql" component={ExampleGQLComponent} />
-        <Route path="/screen-error" component={ScreenErrorComponent} />
       </div>
       <FooterComponent></FooterComponent>
     </Router>
