+++ src/contact-form/contact-form-style.css
@@ -1,0 +1,62 @@
+/* form contact */
+
+.contact-container {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+}
+
+.form-contact {
+    background-color: #171c24;
+    border-radius: 5px;
+    width: 55%;
+    padding: 20px;
+}
+
+.form-contact p {
+    text-align: justify;
+    color: #ffffff;
+}
+
+.form-contact button {
+    padding: 0 15px;
+    color: #fff;
+    background: #ff3d00;
+    border: none;
+    border-radius: 2px;
+    cursor: pointer;
+    font-size: 17px;
+}
+
+.form-contact button:hover {
+    background: #e63600;
+}
+
+.form-contact .icon_input {
+    float: right;
+    margin-right: 6px;
+    margin-top: -34px;
+    position: relative;
+    z-index: 2;
+    color: #171c24;
+}
+
+.form-contact input, .form-contact textarea {
+    width: 100%;
+    padding: 0 10px;
+    margin: 10px 0;
+    border-radius: 5px;
+    background: #ffffff;
+    border: none;
+    font-family: 'Grandstander', cursive;
+}
+
+.form-contact input, .form-contact button {
+    height: 30px;
+}
+
+@media only screen and (min-width:0px) and (max-width:750px) {
+    .form-contact {
+        width: 80%;
+    }
+}
