+++ src/contact/contact-style.css
@@ -1,4 +1,4 @@
-/* form contact */
+/* form contact
 
 .contact-container {
   display: flex;
@@ -59,4 +59,4 @@
   .form-contact {
     width: 80%;
   }
-}
+} */
